Move layout code from @yamcs/displays to yamcs-web
