+++ yamcs-web/packages/displays/src/index.ts
@@ -1,10 +1,6 @@
 export { Display } from './Display';
 export { DisplayCommunicator } from './DisplayCommunicator';
-export { Coordinates, DisplayFrame } from './DisplayFrame';
 export { DisplayHolder } from './DisplayHolder';
-export { Layout, LayoutListener, LayoutStateListener } from './Layout';
-export { LayoutState } from './LayoutState';
 export { OpenDisplayCommandOptions } from './OpenDisplayCommandOptions';
 export { OpiDisplay } from './opi/OpiDisplay';
 export { UssDisplay } from './uss/UssDisplay';
-
